Project restructurization
