+++ project/src/main/java/com/max/grpc/projects/echo/client/Main.java
@@ -1,5 +1,5 @@
 
-package com.max.grpc.project;
+package com.max.grpc.projects.echo.client;
 
 import io.grpc.StatusRuntimeException;
