+++ project/src/main/java/com/max/grpc/projects/echo/server/Main.java
@@ -1,0 +1,12 @@
+
+package com.max.grpc.projects.echo.server;
+
+import java.io.IOException;
+
+public class Main {
+    public static void main(String[] args) throws IOException, InterruptedException {
+        final EchoServer server = new EchoServer();
+        server.start();
+        server.blockUntilShutdown();
+    }
+}
