+++ protos/src/main/proto/echo/echo.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package max.schemas.echo;
 
 option java_multiple_files = true;
-option java_package = "com.max.grpc.protos";
+option java_package = "com.max.grpc.projects.echo.protos";
 option java_outer_classname = "EchoProto";
 
 service EchoService {
